+++ packages/vscode/src/extension.ts
@@ -1,13 +1,13 @@
 import {ExtensionContext} from 'vscode';
 import {SemanticRewriteCommand} from './capabilities/semanticRewrite';
 import {OpenDataFolderCommand} from './capabilities/debug';
-import {WebAppServer} from './capabilities/server';
+import {WebApp} from './capabilities/web';
 
 export function activate(context: ExtensionContext) {
     context.subscriptions.push(
         new SemanticRewriteCommand(),
         new OpenDataFolderCommand(),
-        new WebAppServer()
+        new WebApp(context)
     );
 }