+++ packages/web/src/components/ClientProvider/index.tsx
@@ -1,40 +1,8 @@
 import {createContext, ReactNode, use, useEffect, useState} from 'react';
 import {Protocol} from '@oniichan/server/protocol';
-import {Client, DirectPort, ExecutionMessage, Port} from '@otakustay/ipc';
+import {Client, DirectPort} from '@otakustay/ipc';
 import LoadingSplash from './LoadingSplash';
-
-function isExecutionMesssage(message: any): message is ExecutionMessage {
-    return 'taskId' in message;
-}
-
-class WebSocketPort implements Port {
-    private readonly socket: WebSocket;
-
-    constructor(socket: WebSocket) {
-        this.socket = socket;
-    }
-
-    send(message: ExecutionMessage): void {
-        this.socket.send(JSON.stringify(message));
-    }
-
-    listen(callback: (message: ExecutionMessage) => void): void {
-        this.socket.addEventListener(
-            'message',
-            (event: MessageEvent<string>) => {
-                try {
-                    const data = JSON.parse(event.data);
-                    if (isExecutionMesssage(data)) {
-                        callback(data);
-                    }
-                }
-                catch {
-                    // Discard incorrect JSON messages
-                }
-            }
-        );
-    }
-}
+import {WebSocketPort, VscodeMessagePort} from './port';
 
 interface ClientContextValue {
     client: Client<Protocol>;
@@ -51,19 +19,29 @@ export default function ClientProvider({children}: Props) {
     const [ipcClient, setIpcClient] = useState<Client<Protocol> | null>(null);
     useEffect(
         () => {
-            const socket = new WebSocket(`ws://${location.host}/gateway`);
-            socket.addEventListener(
-                'open',
-                () => {
-                    const port = new WebSocketPort(socket);
-                    const client = new Client<Protocol>(port);
-                    setIpcClient(client);
-                }
-            );
+            const isVscode = location.protocol === 'vscode-webview:';
+            console.log('isVscode: ', isVscode);
 
-            return () => {
-                socket.close();
-            };
+            if (isVscode) {
+                const port = new VscodeMessagePort();
+                const client = new Client<Protocol>(port);
+                setIpcClient(client);
+            }
+            else {
+                const socket = new WebSocket(`ws://${location.host}/gateway`);
+                socket.addEventListener(
+                    'open',
+                    () => {
+                        const port = new WebSocketPort(socket);
+                        const client = new Client<Protocol>(port);
+                        setIpcClient(client);
+                    }
+                );
+
+                return () => {
+                    socket.close();
+                };
+            }
         },
         []
     );